+++ changelog.md
@@ -8,6 +8,7 @@
 - Migrated to MySQL: case-insensitive database resulting in more accurate matching; the consequence is that all IDs are cooerced to lower-case
 - Use server-side caching and static generation of common pages to speed up initial load times
 - Remove ~6k variables that are not available or useful to researchers
+- Add metadata status indicator dot to collection, dataset, and variable lists
 
 # version 0.1.5